Bit&Build2025/pyTeam_Final-Round_10/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
google-generativeai>=0.8.0
requests>=2.31.0
python-multipart>=0.0.6
aiofiles>=23.2.0
pillow>=10.0.0
tavily-python>=0.5.0
reportlab>=4.0.0
markdown>=3.5.0
tweepy>=4.14.0
